+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## FUTURE
+
+* [ADDED] `getOrFail` method
+* [REMOVED] `bi-logger` dependency
+
 ## 0.7.3
 
 * [FIXED] `config` shell option was being ignored when `parse-pos-args`==false